Decentralize  definitions
Move tool definitions from PostmanAPIServer to their respective tool
handler
classes. Each tool class now implements getToolDefinitions() to expose
its own
interface, following the single responsibility principle. This change:

- Reduces duplication in tool definitions and implementations
- Makes it easier to add and maintain tools
- Improves code organization by co-locating related functionality
- Simplifies the main server class by delegating tool management

The refactoring includes:
- Moving PostmanAPIServer to its own file
- Breaking down setupHandlers into focused methods
- Adding getToolDefinitions() to all tool classes
- Updating server to aggregate tool definitions from handlers
